feat(chart): implementa migração do po-chart-new para po-chart
