--- projects/ui/src/lib/components/po-chart-new/index.ts
@@ -1,6 +1,0 @@
-export * from './po-chart-new.component';
-
-export * from './po-chart-new.module';
-
-/* @ToDo - Ao apagar esse arquivo, lembre também de remover a linha "export * from './po-chart-new/index'" no arquivo index.css, que fica em:
-po-angular\projects\ui\src\lib\components. */
