+++ projects/ui/src/lib/components/po-chart/po-chart-base.component.ts
@@ -1,16 +1,16 @@
-import { EventEmitter, Input, Output, Directive, OnChanges, SimpleChanges } from '@angular/core';
+import { Directive, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import { poLocaleDefault } from '../../services/po-language/po-language.constant';
+import { PoLanguageService } from '../../services/po-language/po-language.service';
+import { PoChartType } from '../po-chart/enums/po-chart-type.enum';
+import { poChartLiteralsDefault } from '../po-chart/interfaces/po-chart-literals-default.interface';
+import { PoChartLiterals } from '../po-chart/interfaces/po-chart-literals.interface';
+import { PoChartOptions } from '../po-chart/interfaces/po-chart-options.interface';
+import { PoChartDataLabel } from '../po-chart/interfaces/po-chart-serie-data-label.interface';
+import { PoChartSerie } from '../po-chart/interfaces/po-chart-serie.interface';
+import { PoPopupAction } from '../po-popup';
 
-import { convertToInt, isTypeof } from '../../utils/util';
-
-import { PoChartContainerSize } from './interfaces/po-chart-container-size.interface';
-import { PoChartType } from './enums/po-chart-type.enum';
-import { PoChartOptions } from './interfaces/po-chart-options.interface';
-import { PoChartSerie } from './interfaces/po-chart-serie.interface';
-import { PoColorService } from '../../services/po-color/po-color.service';
-import { PoChartDataLabel } from './interfaces/po-chart-serie-data-label.interface';
-
-const poChartDefaultHeight = 400;
 const poChartMinHeight = 200;
+const poChartDefaultHeight = 400;
 
 /**
  * @description
@@ -29,131 +29,56 @@ const poChartMinHeight = 200;
  *
  * > Veja nosso [guia de uso para gráficos](/guides/guide-charts) para auxiliar na construção do seu gráfico,
  * informando em qual caso utilizar, o que devemos evitar e boas práticas relacionada a cores.
+ *
+ * #### Tokens customizáveis
+ *
+ * É possível alterar o estilo do componente usando os seguintes tokens (CSS):
+ *
+ * > Para maiores informações, acesse o guia [Personalizando o Tema Padrão com Tokens CSS](https://po-ui.io/guides/theme-customization).
+ *
+ * | Propriedade                              | Descrição                                             | Valor Padrão                                      |
+ * |------------------------------------------|-------------------------------------------------------|---------------------------------------------------|
+ * | **Header (.po-chart-header )**           |                                                       |                                                   |
+ * | `--background-color`                     | Cor de background do cabeçalho                        | `var(--color-neutral-light-00)`                   |
+ * | `--color`                                | Cor da fonte do cabeçalho                             | `var(--color-neutral-dark-70)`                    |
+ * | `--font-family`                          | Família tipográfica usada                             | `var(--font-family-theme)`                        |
+ * | `--font-size-title`                      | Tamanho da fonte                                      | `var(--font-size-default)`                        |
+ * | `--font-size-icons`                      | Tamanho dos ícones                                    | `var(--font-size-md)`                             |
+ * | `--font-weight`                          | Peso da fonte                                         | `var(--font-weight-bold)`                         |
+ * | **Chart (.po-chart)**                    |                                                       |                                                   |
+ * | `--background-color-grid`                | Cor de background                                     | `var(--color-neutral-light-00)`                   |
+ * | `--color-grid`                           | Cor da fonte                                          | `var(--color-neutral-light-20)`                   |
+ * | `--font-family-grid`                     | Família tipográfica usada                             | `var(--font-family-theme)`                        |
+ * | `--font-size-grid`                       | Tamanho da fonte                                      | `var(--font-size-xs)`                             |
+ * | `--font-weight-grid`                     | Peso da fonte                                         | `var(--font-weight-normal)`                       |
+ * | `--color-legend`                         | Cor da fonte da legenda                               | `var(--color-neutral-dark-70)`                    |
+ * | `--border-radius-bar`                    | Contém o valor do raio dos cantos do elemento         | `var(--border-radius-none)`                       |
+ * | `--color-grid-hover`                     | Cor no estado hover                                   | `var(--color-neutral-mid-60)`                     |
  */
 @Directive()
-export abstract class PoChartBaseComponent implements OnChanges {
+export abstract class PoChartBaseComponent implements OnInit {
+  private _literals?: PoChartLiterals;
+  private setHeightGauge = false;
+  private readonly language: string;
+
+  @Input('t-id') id: string = 'myChart';
+
   /** Define o título do gráfico. */
   @Input('p-title') title?: string;
-
-  /**
-   * @optional
-   *
-   * @description
-   *
-   * Evento executado quando o usuário clicar sobre um elemento do gráfico.
-   *
-   * O evento emitirá o seguinte parâmetro:
-   * - *donut* e *pie*: um objeto contendo a categoria e valor da série.
-   * - *area*, *line*, *column* e *bar*: um objeto contendo o nome da série, valor e categoria do eixo do gráfico.
-   */
-  @Output('p-series-click')
-  seriesClick = new EventEmitter<PoChartSerie>();
-
-  /**
-   * @optional
-   *
-   * @description
-   *
-   * Evento executado quando o usuário passar o *mouse* sobre um elemento do gráfico.
-   *
-   * O evento emitirá o seguinte parâmetro de acordo com o tipo de gráfico:
-   * - *donut* e *pie*: um objeto contendo a categoria e valor da série.
-   * - *area*, *line*, *column* e *bar*: um objeto contendo a categoria, valor da série e categoria do eixo do gráfico.
-   */
-  @Output('p-series-hover')
-  seriesHover = new EventEmitter<PoChartSerie>();
-
-  // manipulação das séries tratadas internamente para preservar 'p-series';
-  chartSeries: Array<PoChartSerie> = [];
-  chartType: PoChartType;
-  svgContainerSize: PoChartContainerSize;
-
-  private _options: PoChartOptions;
-  private _categories: Array<string>;
-  private _height: number;
-  private _series: Array<PoChartSerie>;
-  private _type: PoChartType;
-
-  private defaultType: PoChartType;
-
-  /**
-   * @optional
-   *
-   * @description
-   *
-   * Define a altura do gráfico.
-   *
-   * > O valor mínimo aceito nesta propriedade é 200.
-   *
-   * @default `400px`
-   */
-  @Input('p-height') set height(value: number) {
-    const intValue = convertToInt(value);
-    let height: number;
-
-    if (isTypeof(value, 'number')) {
-      height = intValue <= poChartMinHeight ? poChartMinHeight : intValue;
-    } else {
-      height = this.setDefaultHeight();
-    }
-
-    this._height = height;
-
-    this.getSvgContainerSize();
-    this.rebuildComponentRef();
-  }
-
-  get height(): number {
-    return this._height || this.setDefaultHeight();
-  }
-
-  /**
-   * @optional
-   *
-   * @description
-   *
-   * Define o tipo de gráfico.
-   *
-   * É possível também combinar gráficos dos tipos linha e coluna. Para isso, opte pela declaração de `type` conforme a interface `PoChartSerie`.
-   *
-   * > Note que, se houver declaração de tipo de gráfico tanto em `p-type` quanto em `PochartSerie.type`, o valor `{ type }` da primeira série anulará o valor definido em `p-type`.
-   *
-   * Se não passado valor, o padrão será relativo à primeira série passada em `p-series`:
-   * - Se `p-series = [{ data: [1,2,3] }]`: será `PoChartType.Column`.
-   * - Se `p-series = [{ data: 1 }]`: será `PoChartType.Pie`.
-   *
-   * > Veja os valores válidos no *enum* `PoChartType`.
-   */
-  @Input('p-type') set type(value: PoChartType) {
-    // O Valor default definido em `p-series` de acordo com a primeira série passada.
-    this._type = (<any>Object).values(PoChartType).includes(value) ? value : undefined;
-
-    this.rebuildComponentRef();
-  }
-
-  get type(): PoChartType {
-    return this._type;
-  }
 
   /**
    * @description
    *
    * Define os elementos do gráfico que serão criados dinamicamente.
    */
-  @Input('p-series') set series(value: Array<PoChartSerie>) {
-    this._series = value || [];
+  @Input('p-series') series: Array<PoChartSerie>;
 
-    if (Array.isArray(this._series) && this._series.length) {
-      this.setTypeDefault(this._series[0]);
-    } else {
-      this.transformObjectToArrayObject(this.series);
-      this.rebuildComponentRef();
-    }
-  }
-
-  get series() {
-    return this._series;
-  }
+  /**
+   * @description
+   *
+   * Define o valor do gráfico do tipo `Gauge`.
+   */
+  @Input('p-value-gauge-multiple') valueGaugeMultiple?: number;
 
   /**
    * @optional
@@ -166,15 +91,9 @@ export abstract class PoChartBaseComponent implements OnChanges {
    *
    * > Caso não seja especificado um valor para a categoria, será plotado um hífen na categoria referente a cada série.
    */
-  @Input('p-categories') set categories(value: Array<string>) {
-    if (Array.isArray(value)) {
-      this._categories = value;
-    }
-  }
+  @Input('p-categories') categories?: Array<string>;
 
-  get categories() {
-    return this._categories;
-  }
+  @Input('p-custom-actions') customActions?: Array<PoPopupAction>;
 
   /**
    * @optional
@@ -197,19 +116,7 @@ export abstract class PoChartBaseComponent implements OnChanges {
    *  };
    * ```
    */
-  @Input('p-options') set options(value: PoChartOptions) {
-    if (value instanceof Object && !(value instanceof Array)) {
-      this._options = value;
-
-      if (this._options.hasOwnProperty('legend') && typeof this._options.legend === 'boolean') {
-        this.getSvgContainerSize();
-      }
-    }
-  }
-
-  get options() {
-    return this._options;
-  }
+  @Input('p-options') options?: PoChartOptions;
 
   /**
    * @optional
@@ -235,81 +142,129 @@ export abstract class PoChartBaseComponent implements OnChanges {
    */
   @Input('p-data-label') dataLabel?: PoChartDataLabel;
 
-  constructor(protected colorService: PoColorService) {}
-
-  get isTypeCircular() {
-    return this.defaultType === PoChartType.Pie || this.defaultType === PoChartType.Donut;
+  /**
+   * @optional
+   *
+   * @description
+   *
+   * Define a altura do gráfico em px.
+   *
+   * > No caso do tipo `Gauge`, o valor padrão é `300` e esse é seu valor minimo aceito. Nos outros tipos, o valor mínimo aceito nesta propriedade é 200.
+   *
+   * @default `400`
+   */
+  @Input('p-height')
+  set height(value: number) {
+    let heightGauge = null;
+    this.setHeightGauge = true;
+    if (this.type === PoChartType.Gauge) {
+      heightGauge = 300;
+    }
+    this._height = Math.max(value ?? heightGauge ?? poChartDefaultHeight, poChartMinHeight);
   }
 
-  ngOnChanges(changes: SimpleChanges): void {
-    const isArrayOfseries = Array.isArray(this.series) && this.series.length > 0;
+  get height(): number {
+    return this._height;
+  }
 
-    if (
-      (changes.series && isArrayOfseries) ||
-      (changes.type && isArrayOfseries) ||
-      (changes.categories && isArrayOfseries)
-    ) {
-      this.validateSerieAndAddType(this.series);
-    }
+  private _height: number = poChartDefaultHeight;
 
-    if ((changes.type && !this.isTypeCircular) || (changes.categories && !this.isTypeCircular)) {
-      this.svgContainerSize = {
-        ...this.svgContainerSize,
-        axisXLabelWidth: this.calculateAxisXLabelArea()
+  /**
+   * @optional
+   *
+   * @description
+   *
+   * Define o tipo de gráfico.
+   *
+   * É possível também combinar gráficos dos tipos linha e coluna. Para isso, opte pela declaração de `type` conforme a interface `PoChartSerie`.
+   *
+   * > Note que, se houver declaração de tipo de gráfico tanto em `p-type` quanto em `PochartSerie.type`, o valor `{ type }` da primeira série anulará o valor definido em `p-type`.
+   *
+   * Se não passado valor, o padrão será relativo à primeira série passada em `p-series`:
+   * - Se `p-series = [{ data: [1,2,3] }]`: será `PoChartType.Column`.
+   * - Se `p-series = [{ data: 1 }]`: será `PoChartType.Pie`.
+   *
+   * > Veja os valores válidos no *enum* `PoChartType`.
+   */
+  @Input('p-type') type: PoChartType;
+
+  /**
+   * @optional
+   *
+   * @description
+   *
+   * Objeto com as literais usadas no `po-chart`.
+   *
+   * Para utilizar basta passar a literal que deseja customizar:
+   *
+   * ```
+   *  const customLiterals: PoChartLiterals = {
+   *    downloadCSV: 'Obter CSV',
+   *  };
+   * ```
+   *
+   * E para carregar a literal customizada, basta apenas passar o objeto para o componente.
+   *
+   * ```
+   * <po-chart
+   *   [p-literals]="customLiterals">
+   * </po-chart>
+   * ```
+   *
+   * > O objeto padrão de literais será traduzido de acordo com o idioma do
+   * [`PoI18nService`](/documentation/po-i18n) ou do browser.
+   */
+  @Input('p-literals') set literals(value: PoChartLiterals) {
+    if (value instanceof Object && !(value instanceof Array)) {
+      this._literals = {
+        ...poChartLiteralsDefault[poLocaleDefault],
+        ...poChartLiteralsDefault[this.language],
+        ...value
       };
+    } else {
+      this._literals = poChartLiteralsDefault[this.language];
     }
   }
 
-  onSeriesClick(event: any): void {
-    this.seriesClick.emit(event);
+  get literals() {
+    return this._literals || poChartLiteralsDefault[this.language];
   }
 
-  onSeriesHover(event: any): void {
-    this.seriesHover.emit(event);
+  /**
+   * @optional
+   *
+   * @description
+   *
+   * Evento executado quando o usuário clicar sobre um elemento do gráfico.
+   *
+   * O evento emitirá o seguinte parâmetro:
+   * - *donut* e *pie*: um objeto contendo a categoria e valor da série.
+   * - *area*, *line*, *column* e *bar*: um objeto contendo o nome da série, valor e categoria do eixo do gráfico.
+   */
+  @Output('p-series-click')
+  seriesClick: EventEmitter<any> = new EventEmitter<any>();
+
+  /**
+   * @optional
+   *
+   * @description
+   *
+   * Evento executado quando o usuário passar o *mouse* sobre um elemento do gráfico.
+   *
+   * O evento emitirá o seguinte parâmetro de acordo com o tipo de gráfico:
+   * - *donut* e *pie*: um objeto contendo a categoria e valor da série.
+   * - *area*, *line*, *column* e *bar*: um objeto contendo a categoria, valor da série e categoria do eixo do gráfico.
+   */
+  @Output('p-series-hover')
+  seriesHover: EventEmitter<any> = new EventEmitter<any>();
+
+  constructor(languageService: PoLanguageService) {
+    this.language = languageService.getShortLanguage();
   }
 
-  private setDefaultHeight() {
-    return poChartDefaultHeight;
+  ngOnInit(): void {
+    if (this.type === PoChartType.Gauge && !this.setHeightGauge) {
+      this._height = 300;
+    }
   }
-
-  private transformObjectToArrayObject(serie) {
-    this.chartSeries = typeof serie === 'object' && Object.keys(serie).length ? [{ ...serie }] : [];
-  }
-
-  private setTypeDefault(serie: PoChartSerie) {
-    const data = serie.data;
-    const serieType = (<any>Object).values(PoChartType).includes(serie.type) ? serie.type : undefined;
-
-    this.defaultType = serieType ? serieType : Array.isArray(data) ? PoChartType.Column : PoChartType.Pie;
-  }
-
-  private validateSerieAndAddType(series: Array<PoChartSerie>): void {
-    const filteredSeries = series.filter(serie =>
-      this.isTypeCircular ? typeof serie.data === 'number' : Array.isArray(serie.data)
-    );
-
-    this.chartSeries = this.appendType(this.appendColors(filteredSeries));
-  }
-
-  private appendColors(series: Array<PoChartSerie>) {
-    return this.colorService.getColors<PoChartSerie>(series);
-  }
-
-  private appendType(series: Array<PoChartSerie>) {
-    return series.map((serie, index) => {
-      if (index === 0) {
-        this.chartType = (<any>Object).values(PoChartType).includes(serie.type)
-          ? serie.type
-          : this.type || this.defaultType;
-      }
-
-      return { ...serie, type: serie.type || this.chartType };
-    });
-  }
-
-  abstract rebuildComponentRef(): void;
-
-  // válido para gráficos do tipo circular e que será refatorado.
-  protected abstract getSvgContainerSize(): void;
-  protected abstract calculateAxisXLabelArea(): number;
 }
