+++ projects/ui/src/lib/components/po-chart/po-chart.component.html
@@ -1,27 +1,56 @@
-<div #chartWrapper class="po-chart-wrapper" (p-resize-observer)="resizeAction()">
-  <div #chartHeader class="po-chart-header">
-    <div class="po-chart-title">{{ title }}</div>
+<div *ngIf="series.length" [class.expanded]="isExpanded" id="chart-container">
+  <div *ngIf="showHeader" class="po-chart-header">
+    <div
+      class="po-chart-header-title po-lg-9 po-md-8 po-sm-6"
+      [p-tooltip]="tooltipTitle"
+      (mouseover)="showTooltipTitle($event)"
+      (focus)="showTooltipTitle($event)"
+    >
+      {{ title }}
+    </div>
+
+    <div class="po-chart-header-actions po-lg-3 po-md-4 po-sm-6">
+      <po-button
+        *ngIf="!options?.header?.hideTableDetails"
+        p-icon="ICON_LIST_BULLETS"
+        p-kind="tertiary"
+        (p-click)="openModal()"
+      ></po-button>
+      <po-button
+        *ngIf="!options?.header?.hideExpand"
+        p-icon="ICON_EXPAND"
+        p-kind="tertiary"
+        (p-click)="toggleExpand()"
+      ></po-button>
+      <po-button
+        *ngIf="showPopup"
+        #targetPopup
+        p-icon="ICON_MORE_VERT"
+        p-kind="tertiary"
+        (p-click)="popup.toggle()"
+      ></po-button>
+    </div>
   </div>
-
-  <po-chart-container
-    [p-options]="options"
-    [p-type]="chartType"
-    [p-series]="chartSeries"
-    [p-categories]="categories"
-    [p-container-size]="svgContainerSize"
-    [p-data-label]="dataLabel"
-    (p-serie-click)="onSeriesClick($event)"
-    (p-serie-hover)="onSeriesHover($event)"
-  ></po-chart-container>
-
-  <!-- Injeção de gráficos do tipo gauge. Remover na depreciação.  -->
-  <ng-template #chartContainer></ng-template>
-
-  <div *ngIf="options?.legend !== false">
-    <ng-container *ngTemplateOutlet="chartLegendGroup"></ng-container>
-  </div>
+  <div
+    id="chart-id"
+    class="po-chart"
+    [ngStyle]="
+      showHeader
+        ? { height: height - headerHeight + 'px', 'margin-top': chartMarginTop }
+        : { height: height + 'px', 'margin-top': chartMarginTop }
+    "
+  ></div>
+  <div
+    #tooltipElement
+    id="custom-tooltip"
+    [p-tooltip]="tooltipText"
+    [p-hide-arrow]="true"
+    [p-inner-html]="true"
+    [p-tooltip-position]="positionTooltip"
+  ></div>
+  <po-popup #popup [p-actions]="popupActions" p-position="bottom" [p-target]="targetRef"> </po-popup>
 </div>
 
-<ng-template #chartLegendGroup>
-  <po-chart-legend #chartLegend [p-series]="chartSeries" [p-type]="type"> </po-chart-legend>
-</ng-template>
+<po-modal #modalComponent [p-click-out]="true" [p-title]="title" [p-primary-action]="actionModal">
+  <po-table [p-hide-columns-manager]="true" [p-items]="itemsTable" [p-columns]="columnsTable"> </po-table>
+</po-modal>
