+++ projects/ui/src/lib/components/po-gauge/po-gauge.component.html
@@ -1,8 +1,8 @@
-<po-chart-new
+<po-chart
   [p-height]="height"
   [p-options]="options"
   [p-series]="series"
   [p-title]="title"
   [p-type]="'gauge'"
   [p-value-gauge-multiple]="valuesMultiple"
-></po-chart-new>
+></po-chart>
